Improve teacher profile and availability handling
Changed TeacherProfile.preferred_distance to DecimalField for more precise distance values. Updated TeacherProfileSerializer to prevent duplicate profiles only on creation. Modified AvailabilityViewSet to delete existing availabilities before creating new ones. Enhanced TeacherProfileViewSet to ensure users can only update their own profiles.

+++ base/models.py
@@ -122,7 +122,7 @@ class TeacherProfile(models.Model):
     
     gender = models.CharField(max_length=20, choices=GENDER_CHOICES, blank=True)
     teaching_mode = models.CharField(max_length=20, choices=TEACHING_CHOICES, blank=True)
-    preferred_distance = models.PositiveIntegerField(default=0, help_text="Preferred distance for teaching in kilometers")
+    preferred_distance = models.DecimalField(max_digits=5, decimal_places=2, default=0, help_text="Preferred distance for teaching in kilometers")
 
 
     def __str__(self):
